backend/requirements.txt
fastapi
uvicorn
pandas
numpy
python-multipart
pydantic
python-dotenv
psycopg2-binary
pdfplumber
cryptography
langchain-google-genai
langchain
langgraph
jupyter
ipykernel
fpdf
pytest
sqlalchemy
